refactor: split app component, app error boundary
- split big `App` component into smaller components, hooks
- create abstract `AppErrorBoundary` component which is used for wrapping table + the whole app, and can be reused
